Fix another -Wall artifact

+++ src/util.c
@@ -177,7 +177,7 @@ int char_width(char input) {
         4, 7, 6, 6, 6, 6, 5, 6, 6, 2, 5, 5, 2, 9, 6, 6,
         6, 6, 6, 6, 5, 6, 6, 6, 6, 6, 6, 4, 2, 5, 7, 0
     };
-    return lookup[input];
+    return lookup[(size_t)input];
 }
 
 int string_width(const char *input) {
